fix: RMSNorm dtype — compute in float32, cast output back to input dtype
F.rms_norm upcasts internally to float32 but does not cast back,
causing float32 output to flow into bf16 SSM linear layers.
Follow the LLaMA pattern: normalize in float32, cast output to
input dtype before weight multiply.

+++ src/chaoscontrol/core.py
@@ -16,8 +16,8 @@ class RMSNorm(nn.Module):
         self.eps = eps
 
     def forward(self, x: torch.Tensor) -> torch.Tensor:
-        normed = F.rms_norm(x, (x.size(-1),), eps=self.eps)
-        return normed * self.weight.to(dtype=x.dtype)
+        normed = F.rms_norm(x.float(), (x.size(-1),), eps=self.eps)
+        return normed.to(x.dtype) * self.weight
 
 
 class FeedForward(nn.Module):
